plugin-projects: invert the inbox template dependency (public inbox cannot depend on private projects)
check-public-dependencies failed: publishable @dxos/plugin-inbox depended on private
@dxos/plugin-projects. The inboxResearch template and the mailbox 'Set up project'
action move into plugin-projects (private -> public is allowed); plugin-inbox reverts
to no plugin-projects dependency. Changeset renamed to @dxos/plugin-routine.
