+++ packages/plugins/plugin-inbox/package.json
@@ -164,7 +164,6 @@
     "@dxos/plugin-markdown": "workspace:*",
     "@dxos/plugin-observability": "workspace:*",
     "@dxos/plugin-preview": "workspace:*",
-    "@dxos/plugin-projects": "workspace:*",
     "@dxos/plugin-routine": "workspace:*",
     "@dxos/plugin-settings": "workspace:*",
     "@dxos/plugin-space": "workspace:*",
