--- packages/plugins/plugin-inbox/src/templates/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-export * from './inbox-research';
