+++ packages/plugins/plugin-inbox/src/translations.ts
@@ -65,7 +65,6 @@ export const translations = [
 
         'sync-mailbox.label': 'Sync',
         'empty-mailbox.message': 'Mailbox empty',
-        'setup-project.label': 'Set up project',
 
         'sync-calendar.label': 'Sync calendar',
         'empty-calendar.message': 'Calendar empty',
