+++ packages/plugins/plugin-inbox/tsconfig.json
@@ -206,9 +206,6 @@
       "path": "../plugin-preview"
     },
     {
-      "path": "../plugin-projects"
-    },
-    {
       "path": "../plugin-routine"
     },
     {
