+++ packages/plugins/plugin-projects/src/capabilities/app-graph-builder.ts
@@ -6,15 +6,18 @@ import * as Effect from 'effect/Effect';
 import * as Option from 'effect/Option';
 
 import { Capability } from '@dxos/app-framework';
-import { AppCapabilities, AppNode, AppNodeMatcher, GraphPath, TypeSection } from '@dxos/app-toolkit';
+import { AppCapabilities, AppNode, AppNodeMatcher, GraphPath, LayoutOperation, TypeSection } from '@dxos/app-toolkit';
 import { Chat } from '@dxos/assistant-toolkit';
 import { Operation, Project } from '@dxos/compute';
 import { Filter, Obj, Query, Type } from '@dxos/echo';
 import { GraphBuilder, Node } from '@dxos/plugin-graph';
+import { Mailbox } from '@dxos/plugin-inbox';
 import { SpaceOperation } from '@dxos/plugin-space';
 
 import { meta } from '#meta';
 import { ProjectOperation } from '#types';
+
+import { inboxResearch } from '../templates';
 
 /**
  * Surfaces all `Project` objects in a space as a sidebar section nested under the assistant (AI) group —
@@ -44,13 +47,57 @@ export default Capability.makeModule(
 
     const actionExtensions = yield* createProjectActionExtension();
     const chatExtensions = yield* createProjectChatsExtension();
+    const mailboxExtensions = yield* createMailboxProjectExtension();
     return Capability.contributes(AppCapabilities.AppGraphBuilder, [
       ...sectionExtensions,
       ...actionExtensions,
       ...chatExtensions,
+      ...mailboxExtensions,
     ]);
   }),
 );
+
+/**
+ * "Set up project" on the primary mailbox node (`systemTag: 'inbox'`, so it appears once, not on
+ * every sibling filter view): creates an Inbox Research project pre-wired to the mailbox (standing
+ * context, inbox/table skills, starter sender-ledger routine) and opens it. Lives here rather than
+ * in plugin-inbox because the template does (see `templates/index.ts` on the public/private
+ * dependency direction); injecting actions into another plugin's nodes is the established pattern
+ * (e.g. plugin-brain's mailbox Analyze action).
+ */
+export const createMailboxProjectExtension = () =>
+  GraphBuilder.createExtension({
+    id: 'mailboxProjectActions',
+    match: (node) =>
+      node.properties.systemTag === 'inbox' && Mailbox.instanceOf(node.data) ? Option.some(node.data) : Option.none(),
+    actions: (mailbox) => {
+      const db = Obj.getDatabase(mailbox);
+      if (!db) {
+        return Effect.succeed([]);
+      }
+
+      return Effect.succeed([
+        Node.makeAction({
+          id: 'setupProject',
+          data: () =>
+            Effect.gen(function* () {
+              const { subject } = yield* Operation.invoke(
+                ProjectOperation.Create,
+                { templateId: inboxResearch.id, subject: mailbox },
+                { spaceId: db.spaceId },
+              );
+              yield* Operation.invoke(LayoutOperation.Open, { subject: [...subject] });
+            }),
+          properties: {
+            label: ['setup-project.label', { ns: meta.profile.key }],
+            icon: 'ph--stack-plus--regular',
+            disposition: 'list-item',
+            testId: 'projectsPlugin.setupProject',
+          },
+        }),
+      ]);
+    },
+  });
 
 /**
  * A project's chats as its navtree children. Ownership is the ECHO parent edge (no `Project` schema
