+++ packages/plugins/plugin-projects/src/templates/inbox-research.test.ts
@@ -9,9 +9,9 @@ import { Instructions, Project, Routine, Skill, Trigger } from '@dxos/compute';
 import { Collection, Database, Feed, Filter, Obj } from '@dxos/echo';
 import { EchoTestBuilder } from '@dxos/echo-client/testing';
 import { EffectEx } from '@dxos/effect';
+import { Mailbox } from '@dxos/plugin-inbox';
 import { TagIndex, Text } from '@dxos/schema';
 
-import { Mailbox } from '../types';
 import { inboxResearch } from './inbox-research';
 
 describe('inbox research project template', () => {
