+++ packages/plugins/plugin-projects/tsconfig.json
@@ -88,6 +88,9 @@
       "path": "../plugin-graph"
     },
     {
+      "path": "../plugin-inbox"
+    },
+    {
       "path": "../plugin-routine"
     },
     {
